// app/components/MeasurementForm/types.ts
export interface MeasurementFields {
  gender: "male" | "female";
  armhole: number | null;
  bicep: number | null;
  chest: number | null;
  cuff_width: number | null;
  hip: number | null;
  kameez_back_length: number | null;
  kameez_front_length: number | null;
  kameez_length: number | null;
  neck_circumference: number | null;
  neck_depth_back: number | null;
  neck_depth_front: number | null;
  shoulder_width: number | null;
  side_slit_length: number | null;
  sleeve_length: number | null;
  waist: number | null;

  ankle_width: number | null;
  crotch_depth: number | null;
  flare_width: number | null;
  knee: number | null;
  salwar_hip: number | null;
  salwar_length: number | null;
  salwar_waist: number | null;
  thigh: number | null;

  bust: number | null;
  shoulder_to_bust_point: number | null;
  bust_point_to_bust_point: number | null;
}